+++ src/containers/MyReading.js
@@ -1,36 +1,22 @@
-import React, { useContext } from 'react'
+import React from 'react'
 import Helmet from 'react-helmet'
 import styled from 'styled-components'
-import {
-  Switch,
-  Route,
-  useLocation,
-  matchPath,
-  useHistory,
-} from 'react-router-dom'
+import { Switch, Route, useLocation, matchPath } from 'react-router-dom'
 
 // @twreporter
-import { Title2 } from '@twreporter/react-components/lib/title-bar'
 import { H2 } from '@twreporter/react-components/lib/text/headline'
 import mq from '@twreporter/core/lib/utils/media-query'
 import { colorGrayscale } from '@twreporter/core/lib/constants/color'
-import { TextButton } from '@twreporter/react-components/lib/button'
-import { Arrow } from '@twreporter/react-components/lib/icon'
-import {
-  DesktopAndAbove,
-  TabletAndBelow,
-} from '@twreporter/react-components/lib/rwd'
 
 // constants
 import siteMeta from '../constants/site-meta'
 import routes from '../constants/routes'
 
 // components
-import SavedBookmarksPage from '../components/my-reading/saved-bookmarks'
-import BrowsingHistoryPage from '../components/my-reading/browsing-history'
-
-// context
-import { CoreContext } from '../contexts'
+import SavedBookmarksPage from '../components/my-reading/saved-bookmarks-page'
+import BrowsingHistoryPage from '../components/my-reading/browsing-history-page'
+import SavedBookmarksSection from '../components/my-reading/saved-bookmakrs-section'
+import BrowsingHistorySection from '../components/my-reading/browsing-history-section'
 
 const GridContainer = styled.div`
   width: 100%;
@@ -76,16 +62,8 @@ const BrowsingHistory = styled.div`
   padding-bottom: 24px;
 `
 
-const EmptyBox = styled.div`
-  margin-top: 24px;
-  margin-bottom: 24px;
-  height: 300px;
-`
-
 const MyReadingPage = () => {
   const { pathname } = useLocation()
-  const navigate = useHistory()
-  const { releaseBranch } = useContext(CoreContext)
 
   const getSiteTitle = pathname => {
     if (matchPath(pathname, routes.myReadingPage.savedBookmarksPage.path)) {
@@ -102,80 +80,6 @@ const MyReadingPage = () => {
   const titleText = getSiteTitle(pathname)
   const title = titleText + siteMeta.name.separator + siteMeta.name.full
   const canonical = `${siteMeta.urlOrigin}${pathname}`
-
-  const SavedBookmarksContent = () => {
-    const moreSavedBookmarkBtn = size => {
-      return (
-        <>
-          <DesktopAndAbove>
-            <TextButton
-              text="查看更多"
-              rightIconComponent={<Arrow direction="right" />}
-              onClick={() =>
-                navigate.push(routes.myReadingPage.savedBookmarksPage.path)
-              }
-              size={TextButton.Size.L}
-              releaseBranch={releaseBranch}
-            />
-          </DesktopAndAbove>
-          <TabletAndBelow>
-            <TextButton
-              text="查看更多"
-              rightIconComponent={<Arrow direction="right" />}
-              onClick={() =>
-                navigate.push(routes.myReadingPage.savedBookmarksPage.path)
-              }
-              size={TextButton.Size.S}
-              releaseBranch={releaseBranch}
-            />
-          </TabletAndBelow>
-        </>
-      )
-    }
-    return (
-      <div>
-        <Title2 title={'已收藏'} renderButton={moreSavedBookmarkBtn()} />
-        <EmptyBox>空無一物</EmptyBox>
-      </div>
-    )
-  }
-
-  const BrowsingHistoryContent = () => {
-    const moreHistoryBtn = () => {
-      return (
-        <>
-          <DesktopAndAbove>
-            <TextButton
-              text="查看更多"
-              rightIconComponent={<Arrow direction="right" />}
-              onClick={() =>
-                navigate.push(routes.myReadingPage.browsingHistoryPage.path)
-              }
-              size={TextButton.Size.L}
-              releaseBranch={releaseBranch}
-            />
-          </DesktopAndAbove>
-          <TabletAndBelow>
-            <TextButton
-              text="查看更多"
-              rightIconComponent={<Arrow direction="right" />}
-              onClick={() =>
-                navigate.push(routes.myReadingPage.browsingHistoryPage.path)
-              }
-              size={TextButton.Size.S}
-              releaseBranch={releaseBranch}
-            />
-          </TabletAndBelow>
-        </>
-      )
-    }
-    return (
-      <div>
-        <Title2 title={'造訪紀錄'} renderButton={moreHistoryBtn()} />
-        <EmptyBox>空無一物</EmptyBox>
-      </div>
-    )
-  }
 
   return (
     <div>
@@ -202,10 +106,10 @@ const MyReadingPage = () => {
             <Route exact path={routes.myReadingPage.path}>
               <PageTitle text={'我的閱讀'} />
               <SavedBookmarks>
-                <SavedBookmarksContent />
+                <SavedBookmarksSection />
               </SavedBookmarks>
               <BrowsingHistory>
-                <BrowsingHistoryContent />
+                <BrowsingHistorySection />
               </BrowsingHistory>
             </Route>
             <Route path={routes.myReadingPage.savedBookmarksPage.path}>